Added host name parsing from the client, so that a hostname, rather than an IP address will be sent in http-connect. This can be expanded for the other types of relays. In order to use the feature, one should add the directive 'parse_host' to the configuration file. Its value can be either "sni" or "http_host".
For "sni", the client is expected to send an SSL client hello with the SNI
extenion containing the server name.
For "http_host" the clien is expected to send an HTTP request with the Host
header.

If the hostname is received, it will be placed (including the port to which
the connection was intended) in the CONNECT method that is sent to the proxy.