Add hashtable management and lookup
Entries written to the designated input pipe are parsed as if it were a hosts file,
although only the first hostname per line is used.

As long as the fifo is held open, the contents of the mapping will remain intact.
Should the fifo be closed, the mapping will be cleared. This allows simple management
of the size of the hostname cache, simply by closing and reopening the fifo periodically.

A sample rdns config stanza would look something like:

rdns {
	fifo = "/tmp/redsocks_dns.fifo";
}