Rename Linux filtering method to netfilter
We are using SO_ORIGINAL_DST for getting the destination address on
Linux. Since it is defined by netfilter subsystem it is common to both
iptables and the newer nftables.

Let's rename define and the method name to denote we are using netfilter
API and then provide handy aliases for "redirector" backwards
compatibility.